crosscluster/logical: break testutils dependency
Release note: None
