+++ pkg/BUILD.bazel
@@ -173,8 +173,10 @@ ALL_TESTS = [
     "//pkg/config/zonepb:zonepb_test",
     "//pkg/config:config_disallowed_imports_test",
     "//pkg/config:config_test",
+    "//pkg/crosscluster/logical:logical_disallowed_imports_test",
     "//pkg/crosscluster/logical:logical_test",
     "//pkg/crosscluster/physical:physical_test",
+    "//pkg/crosscluster/producer:producer_disallowed_imports_test",
     "//pkg/crosscluster/producer:producer_test",
     "//pkg/crosscluster/replicationutils:replicationutils_test",
     "//pkg/crosscluster/streamclient/randclient:randclient_test",
