+++ pkg/crosscluster/replicationtestutils/testutils.go
@@ -22,13 +22,18 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvpb"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts/ptpb"
 	"github.com/cockroachdb/cockroach/pkg/multitenant/mtinfopb"
 	"github.com/cockroachdb/cockroach/pkg/multitenant/tenantcapabilities"
+	"github.com/cockroachdb/cockroach/pkg/repstream/streampb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/security/username"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
+	"github.com/cockroachdb/cockroach/pkg/sql/isql"
 	"github.com/cockroachdb/cockroach/pkg/storage"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/jobutils"
@@ -45,6 +50,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 	"github.com/cockroachdb/cockroach/pkg/util/retry"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 	"github.com/cockroachdb/errors"
 	"github.com/lib/pq"
 	"github.com/stretchr/testify/require"
@@ -319,7 +325,7 @@ func (c *TenantStreamingClusters) Cutover(
 		require.Equal(c.T, cutoverTime, cutoverOutput.GoTime())
 	}
 
-	protectedTimestamp := replicationutils.TestingGetPTSFromReplicationJob(c.T, ctx, c.SrcSysSQL, c.SrcSysServer, jobspb.JobID(producerJobID))
+	protectedTimestamp := TestingGetPTSFromReplicationJob(c.T, ctx, c.SrcSysSQL, c.SrcSysServer, jobspb.JobID(producerJobID))
 	require.LessOrEqual(c.T, protectedTimestamp.GoTime(), cutoverOutput.GoTime())
 
 	// PTS should be less than or equal to retained time as a result of heartbeats.
@@ -710,7 +716,7 @@ func GetStreamJobIds(
 		destTenantName).Scan(&tenantInfoBytes)
 	require.NoError(t, protoutil.Unmarshal(tenantInfoBytes, &tenantInfo))
 
-	stats := replicationutils.TestingGetStreamIngestionStatsFromReplicationJob(t, ctx, sqlRunner, int(tenantInfo.PhysicalReplicationConsumerJobID))
+	stats := TestingGetStreamIngestionStatsFromReplicationJob(t, ctx, sqlRunner, int(tenantInfo.PhysicalReplicationConsumerJobID))
 	return int(stats.IngestionDetails.StreamID), int(tenantInfo.PhysicalReplicationConsumerJobID)
 }
 
@@ -745,4 +751,91 @@ func SSTMaker(t *testing.T, keyValues []roachpb.KeyValue) kvpb.RangeFeedSSTable 
 
 func WaitForAllProducerJobsToFail(t *testing.T, sql *sqlutils.SQLRunner) {
 	sql.CheckQueryResultsRetry(t, "SELECT distinct(status) FROM [SHOW JOBS] where job_type = 'REPLICATION STREAM PRODUCER'", [][]string{{"failed"}})
+}
+
+func WaitForPTSProtection(
+	t *testing.T,
+	ctx context.Context,
+	sqlRunner *sqlutils.SQLRunner,
+	srv serverutils.ApplicationLayerInterface,
+	producerJobID jobspb.JobID,
+	minTime hlc.Timestamp,
+) {
+	testutils.SucceedsSoon(t, func() error {
+		protected := TestingGetPTSFromReplicationJob(t, ctx, sqlRunner, srv, producerJobID)
+		if protected.Less(minTime) {
+			return errors.Newf("pts %s is less than min time %s", protected, minTime)
+		}
+		return nil
+	})
+}
+
+func WaitForPTSProtectionToNotExist(
+	t *testing.T,
+	ctx context.Context,
+	sqlRunner *sqlutils.SQLRunner,
+	srv serverutils.ApplicationLayerInterface,
+	producerJobID jobspb.JobID,
+) {
+	ptsRecordID := getPTSRecordIDFromProducerJob(t, sqlRunner, producerJobID)
+	ptsProvider := srv.ExecutorConfig().(sql.ExecutorConfig).ProtectedTimestampProvider
+	testutils.SucceedsSoon(t, func() error {
+		err := srv.InternalDB().(descs.DB).Txn(ctx, func(ctx context.Context, txn isql.Txn) error {
+			_, err := ptsProvider.WithTxn(txn).GetRecord(ctx, ptsRecordID)
+			return err
+		})
+		if errors.Is(err, protectedts.ErrNotExists) {
+			return nil
+		}
+		if err == nil {
+			return errors.New("PTS record still exists")
+		}
+		return err
+	})
+}
+
+func getPTSRecordIDFromProducerJob(
+	t *testing.T, sqlRunner *sqlutils.SQLRunner, producerJobID jobspb.JobID,
+) uuid.UUID {
+	payload := jobutils.GetJobPayload(t, sqlRunner, producerJobID)
+	return payload.GetStreamReplication().ProtectedTimestampRecordID
+}
+
+func TestingGetPTSFromReplicationJob(
+	t *testing.T,
+	ctx context.Context,
+	sqlRunner *sqlutils.SQLRunner,
+	srv serverutils.ApplicationLayerInterface,
+	producerJobID jobspb.JobID,
+) hlc.Timestamp {
+	ptsRecordID := getPTSRecordIDFromProducerJob(t, sqlRunner, producerJobID)
+	ptsProvider := srv.ExecutorConfig().(sql.ExecutorConfig).ProtectedTimestampProvider
+
+	var ptsRecord *ptpb.Record
+	err := srv.InternalDB().(descs.DB).Txn(ctx, func(ctx context.Context, txn isql.Txn) error {
+		var err error
+		ptsRecord, err = ptsProvider.WithTxn(txn).GetRecord(ctx, ptsRecordID)
+		return err
+	})
+	require.NoError(t, err)
+
+	return ptsRecord.Timestamp
+}
+
+func TestingGetStreamIngestionStatsFromReplicationJob(
+	t *testing.T, ctx context.Context, sqlRunner *sqlutils.SQLRunner, ingestionJobID int,
+) *streampb.StreamIngestionStats {
+	payload := jobutils.GetJobPayload(t, sqlRunner, jobspb.JobID(ingestionJobID))
+	progress := jobutils.GetJobProgress(t, sqlRunner, jobspb.JobID(ingestionJobID))
+	details := payload.GetStreamIngestion()
+	stats, err := replicationutils.GetStreamIngestionStats(ctx, *details, *progress)
+	require.NoError(t, err)
+	return stats
+}
+
+func GetProducerJobIDFromLDRJob(
+	t *testing.T, sqlRunner *sqlutils.SQLRunner, ldrJobID jobspb.JobID,
+) jobspb.JobID {
+	payload := jobutils.GetJobPayload(t, sqlRunner, ldrJobID)
+	return jobspb.JobID(payload.GetLogicalReplicationDetails().StreamID)
 }
