+++ pkg/crosscluster/replicationutils/BUILD.bazel
@@ -9,8 +9,6 @@ go_library(
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/kv/kvpb",
-        "//pkg/kv/kvserver/protectedts",
-        "//pkg/kv/kvserver/protectedts/ptpb",
         "//pkg/repstream/streampb",
         "//pkg/roachpb",
         "//pkg/sql",
@@ -20,17 +18,11 @@ go_library(
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/isql",
         "//pkg/storage",
-        "//pkg/testutils",
         "//pkg/testutils/fingerprintutils",
-        "//pkg/testutils/jobutils",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
         "//pkg/util/ctxgroup",
         "//pkg/util/hlc",
         "//pkg/util/timeutil",
-        "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
-        "@com_github_stretchr_testify//require",
     ],
 )
