+++ pkg/testutils/jobutils/BUILD.bazel
@@ -14,9 +14,9 @@ go_library(
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/security/username",
-        "//pkg/testutils",
         "//pkg/testutils/sqlutils",
         "//pkg/util/protoutil",
+        "//pkg/util/retry",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_stretchr_testify//require",
     ],
